python-python/pyproject.toml
[project]
name = "grpc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio-tools>=1.69.0",
    "grpcio>=1.69.0",
]
